Bump Python requirement to 3.10.
