+++ .github/workflows/python_tests.yml
@@ -84,7 +84,7 @@ jobs:
     strategy:
       matrix:
         os: [macos-13, windows-latest]
-        python-version: ["3.8"]
+        python-version: ["3.10"]
     steps:
     - uses: actions/checkout@a5ac7e51b41094c92402da3b24376905380afc29 # v4.1.6
       with:
@@ -174,7 +174,7 @@ jobs:
     strategy:
       matrix:
         config:
-          - {os: windows-latest, python-version: '3.8'}
+          - {os: windows-latest, python-version: '3.10'}
 
     steps:
     - uses: actions/checkout@a5ac7e51b41094c92402da3b24376905380afc29 # v4.1.6
@@ -218,7 +218,7 @@ jobs:
     strategy:
       matrix:
         config:
-          - {os: ubuntu-latest, python-version: "3.8"}
+          - {os: ubuntu-latest, python-version: "3.10"}
 
     steps:
     - uses: actions/checkout@a5ac7e51b41094c92402da3b24376905380afc29 # v4.1.6
@@ -271,7 +271,7 @@ jobs:
     strategy:
       matrix:
         config:
-          - {os: ubuntu-latest, python-version: "3.8"}
+          - {os: ubuntu-latest, python-version: "3.10"}
 
     steps:
     - uses: actions/checkout@a5ac7e51b41094c92402da3b24376905380afc29 # v4.1.6
@@ -318,10 +318,10 @@ jobs:
         with:
           submodules: 'true'
 
-      - name: Set up Python 3.8
+      - name: Set up Python 3.10
         uses: actions/setup-python@39cd14951b08e74b54015e9e001cdefcf80e669f # v5.1.1
         with:
-          python-version: 3.8
+          python-version: "3.10"
 
       - name: Install ninja
         run: |
