+++ .github/workflows/r_tests.yml
@@ -86,7 +86,7 @@ jobs:
 
     - uses: actions/setup-python@39cd14951b08e74b54015e9e001cdefcf80e669f # v5.1.1
       with:
-        python-version: "3.8"
+        python-version: "3.10"
         architecture: 'x64'
 
     - uses: r-lib/actions/setup-tinytex@v2
