+++ demo/nvflare/vertical/README.md
@@ -8,7 +8,7 @@ This directory contains a demo of Vertical Federated Learning using
 To run the demo, first build XGBoost with the federated learning plugin enabled (see the
 [README](../../../plugin/federated/README.md)).
 
-Install NVFlare (note that currently NVFlare only supports Python 3.8):
+Install NVFlare:
 ```shell
 pip install nvflare
 ```
