+++ tests/ci_build/conda_env/python_lint.yml
@@ -2,11 +2,11 @@ name: python_lint
 channels:
 - conda-forge
 dependencies:
-- python=3.8
+- python=3.10
 - pylint<3.2.4  # https://github.com/pylint-dev/pylint/issues/9751
 - wheel
 - setuptools
-- mypy>=0.981
+- mypy
 - numpy
 - scipy
 - pandas
